requirements.txt
streamlit>=1.47.0,<2.0.0
pandas>=2.3.1
beautifulsoup4>=4.13.4
openpyxl>=3.1.5
PyPDF2>=3.0.1
psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.41
alembic>=1.16.4
python-dateutil>=2.8.0
python-dotenv>=1.0.0
